pyproject.toml
[tool.poetry]
name = "Vacansy_search_with_PSQL"
version = "0.1.0"
description = ""
authors = ["Gordey <gordej.taran@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"
pandas = "^2.2.3"
ticker = "^0.13"
CacheControl="0.14.2"
certifi="2025.1.31"
charset-normalizer="3.4.1"
cleo="2.1.0"
colorama="0.4.6"
coverage="7.7.1"
crashtest="0.4.1"
distlib="0.3.9"
dulwich="0.22.8"
fastjsonschema="2.21.1"
filelock="3.17.0"
findpython="0.6.3"
flake8="7.1.2"
h11="0.14.0"
httpcore="1.0.7"
httpx="0.28.1"
idna="3.10"
iniconfig="2.0.0"
installer="0.7.0"
jaraco.classes="3.4.0"
jaraco.context="6.0.1"
jaraco.functools="4.1.0"
keyring="25.6.0"
mccabe="0.7.0"
more-itertools="10.6.0"
msgpack="1.1.0"
packaging="24.2"
pbs-installer="2025.2.12"
pkginfo="1.12.1.2"
platformdirs="4.3.6"
pluggy="1.5.0"
poetry="2.1.1"
poetry-core="2.1.1"
psycopg2="2.9.10"
pycodestyle="2.12.1"
pyflakes="3.2.0"
pyproject_hooks="1.2.0"
pytest="8.3.5"
pytest-cov="6.0.0"
pywin32-ctypes="0.2.3"
RapidFuzz="3.12.2"
requests="2.32.3"
requests-toolbelt="1.0.0"
shellingham="1.5.4"
sniffio="1.3.1"
tomlkit="0.13.2"
trove-classifiers="2025.3.3.18"
urllib3="2.3.0"
virtualenv="20.29.3"
zstandard="0.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
